requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
passlib
python-jose
